refactor: Move some Ktor utils to a dedicated shared module
Move some Ktor utility code to a new shared `ktor-utils` module to make
it available to components. This commit only moves code which will be
required in upcoming commits, more code will be moved later on demand.
