+++ core/src/main/kotlin/plugins/StatusPages.kt
@@ -30,7 +30,6 @@ import io.ktor.server.response.respond
 import org.eclipse.apoapsis.ortserver.core.api.AuthenticationException
 import org.eclipse.apoapsis.ortserver.core.api.AuthorizationException
 import org.eclipse.apoapsis.ortserver.core.api.ErrorResponse
-import org.eclipse.apoapsis.ortserver.core.utils.UrlPathFormatException
 import org.eclipse.apoapsis.ortserver.dao.QueryParametersException
 import org.eclipse.apoapsis.ortserver.dao.UniqueConstraintException
 import org.eclipse.apoapsis.ortserver.services.InvalidSecretReferenceException
@@ -38,6 +37,7 @@ import org.eclipse.apoapsis.ortserver.services.OrganizationNotEmptyException
 import org.eclipse.apoapsis.ortserver.services.ReferencedEntityException
 import org.eclipse.apoapsis.ortserver.services.ReportNotFoundException
 import org.eclipse.apoapsis.ortserver.services.ResourceNotFoundException
+import org.eclipse.apoapsis.ortserver.shared.ktorutils.UrlPathFormatException
 
 import org.jetbrains.exposed.dao.exceptions.EntityNotFoundException
