+++ settings.gradle.kts
@@ -54,6 +54,7 @@ include(":services:hierarchy")
 include(":services:infrastructure")
 include(":services:report-storage")
 include(":services:secret")
+include(":shared:ktor-utils")
 include(":storage")
 include(":storage:azure-blob")
 include(":storage:database")
