+++ shared/ktor-utils/src/main/kotlin/Extensions.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2025 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.eclipse.apoapsis.ortserver.shared.ktorutils
+
+import io.ktor.server.application.ApplicationCall
+
+/**
+ * Get the parameter from this [ApplicationCall].
+ */
+fun ApplicationCall.requireParameter(name: String) = requireNotNull(parameters[name]) {
+    "Parameter '$name' cannot be null."
+}
+
+/**
+ * Get the ID parameter from this [ApplicationCall]. Throw a [UrlPathFormatException] if the parameter is null or not a
+ * valid ID.
+ */
+fun ApplicationCall.requireIdParameter(name: String): Long {
+    val id = requireParameter(name).toLongOrNull()
+
+    return if (id != null && id > 0) id else throw UrlPathFormatException("Invalid entity ID: '${parameters[name]}'.")
+}
